feat(lcu): error display
